Days 03, 04, 06
Playing catch-up a bit here; including my solutions for days 03 and 04
in both TypeScript and Rust. I was having some issues getting my head
wrapped around processing the input of day 05, so I skipped it and did
day 06. Note that day 06 was only completed in TypeScript because I am
presently not able to dedicate as much time as I thought to this
challenge (as is apparent by how far behind I am!) so for now I am going
to complete each day's problems in TypeScript only for now and if I
manage to get all caught up again I will go back and re-do the missing
Rust implementations.

For reference, as of making this commit up through day 10 is available.